Add hacky push-to-talk support.
Because Wayland forbids processes from listening to system-wide
keypresses, Discord and WebCord are unable to listen for a push-to-talk
key. This commit hacks onto WebCord's socket server and allows for
sending a payload to activate and deactivate push-to-talk. The delay
between sending the packet and push-to-talk state changing appears to be
low enough for daily use. You could activate push-to-talk like so:

`echo '{"pttAction": "activate"}' | websocat --origin https://127.0.0.1 ws://127.0.0.1:6463`

The two "actions" that are supported are `activate` and `deactivate`.

The "hack" part of this commit also refers to the activation method -
the F12 key is used to trigger push-to-talk. Make sure that F12 is bound
before triggering actions, otherwise nothing will happen.